fix: APPS_FOR_TESTING moved to test.ts
