--- APPS_FOR_TESTING.ts
@@ -1,6 +1,0 @@
-export const APP_NAMES = {
-  JOB_TEMPLATE_APP: 'testing-with-jobs-app',
-  MINIMAL_SEED_APP: 'testing-with-minimal-seed-app',
-  REACT_SEED_APP: 'testing-with-react-seed-app',
-  ANGULAR_SEED_APP: 'testing-with-angular-seed-app'
-}
