+++ src/commands/compile/spec/compileJobFile.spec.ts
@@ -5,11 +5,11 @@ import { removeFromGlobalConfig } from '../../../utils/config'
 import {
   createTestGlobalTarget,
   resetTestAppAndReuse,
-  verifyCompiledJob
+  verifyCompiledJob,
+  APP_NAMES
 } from '../../../utils/test'
 import { compileJobFile } from '../internal/compileJobFile'
 import { copy, fileExists, createFolder, readFile } from '@sasjs/utils'
-import { APP_NAMES } from '../../../../APPS_FOR_TESTING'
 
 const fakeJobInit = `/**
   @file
