+++ src/utils/test.ts
@@ -33,6 +33,13 @@ import { Folder, File } from '../types'
 import { ServiceConfig } from '@sasjs/utils/types/config'
 import { create } from '../commands/create/create'
 
+export const APP_NAMES = {
+  JOB_TEMPLATE_APP: 'testing-with-jobs-app',
+  MINIMAL_SEED_APP: 'testing-with-minimal-seed-app',
+  REACT_SEED_APP: 'testing-with-react-seed-app',
+  ANGULAR_SEED_APP: 'testing-with-angular-seed-app'
+}
+
 export const createTestApp = async (parentFolder: string, appName: string) => {
   process.projectDir = parentFolder
   await create(appName, '')
